Add cli test
